page api added

+++ admin-app/src/helpers/axios.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import { api } from '../UrlConfig';
 import store from '../store';
-import { authConstants } from '../action';
+import { authConstants } from '../action/constants';
 
 const authtoken = window.localStorage.getItem('authtoken')
 
@@ -23,9 +23,8 @@ axiosInstance.interceptors.request.use((req) =>{
 axiosInstance.interceptors.response.use((res)=>{
     return res
 },(error)=>{
-    console.log(error.response);
-    const {status} = error.response;
-    if(status === 500){
+    const status = error.response ? error.response.status : 500;
+    if(status && status === 500){
         localStorage.clear();
         store.dispatch({type: authConstants.LOGOUT_SUCCESS})
     }
